feat: Add "started_at" field to JobRequest
A new field, "started_at", has been added to the JobRequest in the remote task system. It's included in the insertion and update methods for JobRequest in task_crud_repositories. This will allow tracking of when each job began. Also, log message was added when a new node is created.

Signed-off-by: Piotr Marecki <pmarecki@gmail.com>