feat: Implement job scheduling and completion functions
This update expands the functionality of the job management tasks by adding job scheduling and job completion methods. Additionally, logging info for newly created jobs is included and a new column `started_at` is added to the 'jobrequest' table to track when a job starts. Model upgrades, including necessary import amendments, are also made in this commit.

Signed-off-by: Piotr Marecki <pmarecki@gmail.com>